[untested] added a camera

+++ src/graphics/gl_functions.h
@@ -8,7 +8,7 @@
 namespace graphics {
 	GLFWwindow * createGLFWContext(int width, int height, std::string name);
 	void windowResizeCallback(GLFWwindow* window, int width, int height);
-	void update(GLFWwindow* window, GameObject* mainScene, float lastTime, float check, int fps);
+	void update(GLFWwindow* window, GameObject* mainScene, double lastTime, float check, int fps);
 	void compileShader(std::string shader);
 	void loadShaders(std::vector<std::string> shadersPaths);
 	void updateCamera(GLFWwindow* window);
